+++ apps/astarte_realm_management/lib/astarte_realm_management/engine.ex
@@ -29,6 +29,8 @@ defmodule Astarte.RealmManagement.Engine do
   alias Astarte.Core.Triggers.SimpleTriggersProtobuf.TaggedSimpleTrigger
   alias Astarte.Core.Triggers.SimpleTriggersProtobuf.TriggerTargetContainer
   alias Astarte.Core.Triggers.Trigger
+  alias Astarte.Core.Triggers.Policy
+  alias Astarte.Core.Triggers.PolicyProtobuf.Policy, as: PolicyProto
   alias Astarte.DataAccess.Database
   alias Astarte.DataAccess.Interface
   alias Astarte.DataAccess.Mappings
@@ -500,7 +502,13 @@ defmodule Astarte.RealmManagement.Engine do
     end
   end
 
-  def install_trigger(realm_name, trigger_name, action, serialized_tagged_simple_triggers) do
+  def install_trigger(
+        realm_name,
+        trigger_name,
+        trigger_policy_name,
+        action,
+        serialized_tagged_simple_triggers
+      ) do
     _ =
       Logger.info("Going to install a new trigger.",
         trigger_name: trigger_name,
@@ -511,14 +519,15 @@ defmodule Astarte.RealmManagement.Engine do
          {:exists?, {:error, :trigger_not_found}} <-
            {:exists?, Queries.retrieve_trigger_uuid(client, trigger_name)},
          simple_trigger_maps = build_simple_trigger_maps(serialized_tagged_simple_triggers),
-         trigger = build_trigger(trigger_name, simple_trigger_maps, action),
+         trigger = build_trigger(trigger_name, trigger_policy_name, simple_trigger_maps, action),
          %Trigger{trigger_uuid: trigger_uuid} = trigger,
          {:ok, action_map} <- Jason.decode(action),
          trigger_target = target_from_action(action_map, trigger_uuid),
          t_container = build_trigger_target_container(trigger_target),
          :ok <- validate_simple_triggers(client, simple_trigger_maps),
          # TODO: these should be batched together
-         :ok <- install_simple_triggers(client, simple_trigger_maps, trigger_uuid, t_container) do
+         :ok <- install_simple_triggers(client, simple_trigger_maps, trigger_uuid, t_container),
+         :ok <- install_trigger_policy_link(client, trigger_uuid, trigger_policy_name) do
       _ =
         Logger.info("Installing trigger.",
           trigger_name: trigger_name,
@@ -578,17 +587,25 @@ defmodule Astarte.RealmManagement.Engine do
     end
   end
 
-  defp build_trigger(trigger_name, simple_trigger_maps, action) do
+  defp build_trigger(trigger_name, policy, simple_trigger_maps, action) do
     simple_trigger_uuids =
       for simple_trigger_map <- simple_trigger_maps do
         simple_trigger_map[:simple_trigger_uuid]
+      end
+
+    policy =
+      if policy == "" do
+        nil
+      else
+        policy
       end
 
     %Trigger{
       trigger_uuid: :uuid.get_v4(),
       simple_triggers_uuids: simple_trigger_uuids,
       action: action,
-      name: trigger_name
+      name: trigger_name,
+      policy: policy
     }
   end
 
@@ -700,6 +717,20 @@ defmodule Astarte.RealmManagement.Engine do
     end)
   end
 
+  defp install_trigger_policy_link(_client, _trigger_uuid, nil) do
+    :ok
+  end
+
+  defp install_trigger_policy_link(_client, _trigger_uuid, "") do
+    :ok
+  end
+
+  defp install_trigger_policy_link(client, trigger_uuid, trigger_policy_name) do
+    with :ok <- verify_trigger_policy_exists(client, trigger_policy_name) do
+      Queries.install_trigger_policy_link(client, trigger_uuid, trigger_policy_name)
+    end
+  end
+
   def get_trigger(realm_name, trigger_name) do
     _ = Logger.debug("Get trigger.", trigger_name: trigger_name)
 
@@ -762,12 +793,15 @@ defmodule Astarte.RealmManagement.Engine do
       _ =
         Logger.info("Deleting trigger.", trigger_name: trigger_name, tag: "delete_trigger_started")
 
-      delete_all_succeeded =
+      delete_all_simple_triggers_succeeded =
         Enum.all?(trigger.simple_triggers_uuids, fn simple_trigger_uuid ->
           Queries.delete_simple_trigger(client, trigger.trigger_uuid, simple_trigger_uuid) == :ok
         end)
 
-      if delete_all_succeeded do
+      delete_policy_link_succeeded =
+        Queries.delete_trigger_policy_link(client, trigger.trigger_uuid, trigger.policy) == :ok
+
+      if delete_all_simple_triggers_succeeded and delete_policy_link_succeeded do
         Queries.delete_trigger(client, trigger_name)
       else
         Logger.warn("Failed to delete trigger.",
@@ -789,5 +823,170 @@ defmodule Astarte.RealmManagement.Engine do
       Config.cassandra_node!(),
       cqex_options
     )
+  end
+
+  def install_trigger_policy(realm_name, policy_json, opts \\ []) do
+    _ = Logger.info("Going to install a new trigger policy.", tag: "install_trigger_policy")
+
+    with {:ok, client} <- connect_to_db_with_realm(realm_name),
+         {:ok, json_obj} <- decode_policy(policy_json),
+         policy_changeset = Policy.changeset(%Policy{}, json_obj),
+         {:ok, %Policy{name: policy_name} = policy} <- validate_trigger_policy(policy_changeset),
+         :ok <- verify_trigger_policy_not_exists(client, policy_name) do
+      _ =
+        Logger.info("Installing trigger policy",
+          tag: "install_policy_started",
+          policy_name: policy_name
+        )
+
+      policy_proto =
+        policy
+        |> Policy.to_policy_proto()
+        |> PolicyProto.encode()
+
+      if opts[:async] do
+        Task.start(Queries, :install_new_trigger_policy, [client, policy_name, policy_proto])
+
+        {:ok, :started}
+      else
+        Queries.install_new_trigger_policy(client, policy_name, policy_proto)
+      end
+    end
+  end
+
+  defp validate_trigger_policy(policy_changeset) do
+    with {:error, %Ecto.Changeset{} = changeset} <-
+           Ecto.Changeset.apply_action(policy_changeset, :insert) do
+      _ =
+        Logger.warn("Received invalid trigger policy: #{inspect(changeset)}.",
+          tag: "invalid_trigger_policy"
+        )
+
+      {:error, :invalid_trigger_policy}
+    end
+  end
+
+  defp decode_policy(policy_json) do
+    with {:error, {:invalid, _invalid_str, _invalid_pos}} <- Jason.decode(policy_json) do
+      _ =
+        Logger.warn("Received invalid trigger policy JSON: #{inspect(policy_json)}.",
+          tag: "invalid_trigger_policy_json"
+        )
+
+      {:error, :invalid_trigger_policy_json}
+    end
+  end
+
+  def get_trigger_policies_list(realm_name) do
+    _ = Logger.debug("Get trigger policy list", tag: "get_trigger_policies_list")
+
+    with {:ok, client} <- connect_to_db_with_realm(realm_name) do
+      Queries.get_trigger_policies_list(client)
+    end
+  end
+
+  def trigger_policy_source(realm_name, policy_name) do
+    _ =
+      Logger.debug("Get trigger policy source.",
+        tag: "trigger_policy_source",
+        policy_name: policy_name
+      )
+
+    with {:ok, client} <- connect_to_db_with_realm(realm_name),
+         {:ok, policy_proto} <- fetch_trigger_policy(client, policy_name) do
+      policy_proto
+      |> PolicyProto.decode()
+      |> Policy.from_policy_proto!()
+      |> Jason.encode()
+    end
+  end
+
+  defp fetch_trigger_policy(client, policy_name) do
+    with {:error, :policy_not_found} <- Queries.fetch_trigger_policy(client, policy_name) do
+      Logger.warn("Trigger policy not found",
+        tag: "trigger_policy_not_found",
+        policy_name: policy_name
+      )
+
+      {:error, :trigger_policy_not_found}
+    end
+  end
+
+  def delete_trigger_policy(realm_name, policy_name, opts \\ []) do
+    _ =
+      Logger.info("Going to delete trigger policy #{policy_name}",
+        tag: "delete_trigger_policy",
+        policy_name: policy_name
+      )
+
+    with {:ok, client} <- connect_to_db_with_realm(realm_name),
+         :ok <- verify_trigger_policy_exists(client, policy_name),
+         {:ok, false} <- check_trigger_policy_has_triggers(client, policy_name) do
+      if opts[:async] do
+        Task.start_link(Engine, :execute_trigger_policy_deletion, [client, policy_name])
+
+        {:ok, :started}
+      else
+        Engine.execute_trigger_policy_deletion(client, policy_name)
+      end
+    end
+  end
+
+  defp verify_trigger_policy_not_exists(client, policy_name) do
+    with {:ok, exists?} <- Queries.check_trigger_policy_already_present(client, policy_name) do
+      if not exists? do
+        :ok
+      else
+        Logger.warn("Trigger policy #{policy_name} already present",
+          tag: "trigger_policy_already_present"
+        )
+
+        {:error, :trigger_policy_already_present}
+      end
+    end
+  end
+
+  defp verify_trigger_policy_exists(client, policy_name) do
+    with {:ok, exists?} <- Queries.check_trigger_policy_already_present(client, policy_name) do
+      if exists? do
+        :ok
+      else
+        Logger.warn("Trigger policy #{policy_name} not found",
+          tag: "trigger_policy_not_found"
+        )
+
+        {:error, :trigger_policy_not_found}
+      end
+    end
+  end
+
+  defp check_trigger_policy_has_triggers(client, policy_name) do
+    with {:ok, true} <- Queries.check_policy_has_triggers(client, policy_name) do
+      Logger.warn("Trigger policy #{policy_name} is currently being used by triggers",
+        tag: "cannot_delete_currently_used_trigger_policy"
+      )
+
+      {:error, :cannot_delete_currently_used_trigger_policy}
+    end
+  end
+
+  def execute_trigger_policy_deletion(client, policy_name) do
+    _ =
+      Logger.info("Trigger policy deletion started.",
+        policy_name: policy_name,
+        tag: "delete_trigger_policy_started"
+      )
+
+    Queries.delete_trigger_policy(client, policy_name)
+  end
+
+  defp connect_to_db_with_realm(realm_name) do
+    with {:error, :database_connection_error} <- Database.connect(realm: realm_name) do
+      Logger.warn("Could not connect to database, realm #{realm_name}",
+        tag: "database_connection_error"
+      )
+
+      {:error, :database_connection_error}
+    end
   end
 end